Move simple send logic to the other logic methods
With the simple send logic, all transaction effects are now executed in tx.cpp. The sending/granting logic was further seperated/refined for the "grant tokens" transaction type.

This is not a functional change.

Move:

- calculateFundraiser (omnicore.cpp -> sp.cpp)
- CMPTransaction::interpretPacket() (omnicore.cpp -> tx.cpp)
- CMPTransaction::logicMath_SimpleSend() (omnicore.cpp -> tx.cpp)

Add helper:

- CMPTransaction::logicHelper_CrowdsaleParticipation()

The helper allows to reuse the crowdsale participation logic, and eventually, if intended, to disable the side effect of crowdsale participations, when granting tokens.
