+++ src/omnicore/omnicore.cpp
@@ -69,7 +69,6 @@
 #include <vector>
 
 using boost::algorithm::token_compress_on;
-using boost::multiprecision::cpp_int;
 using boost::multiprecision::int128_t;
 using boost::to_string;
 
@@ -379,90 +378,6 @@ bool mastercore::update_tally_map(const std::string& who, uint32_t propertyId, i
     }
 
     return bRet;
-}
-
-// calculateFundraiser does token calculations per transaction
-// calcluateFractional does calculations for missed tokens
-static void calculateFundraiser(int64_t amtTransfer, uint8_t bonusPerc,
-        int64_t fundraiserSecs, int64_t currentSecs, int64_t numProps, uint8_t issuerPerc, int64_t totalTokens,
-        std::pair<int64_t, int64_t>& tokens, bool& close_crowdsale)
-{
-    // Weeks in seconds
-    int128_t weeks_sec_(604800);
-
-    // Precision for all non-bitcoin values (bonus percentages, for example)
-    int128_t precision_(1000000000000LL);
-
-    // Precision for all percentages (10/100 = 10%)
-    int128_t percentage_precision(100);
-
-    // Calculate the bonus seconds
-    int128_t bonusSeconds_ = fundraiserSecs - currentSecs;
-
-    // Calculate the whole number of weeks to apply bonus
-    int128_t weeks_ = (bonusSeconds_ / weeks_sec_) * precision_;
-    weeks_ += ((bonusSeconds_ % weeks_sec_) * precision_) / weeks_sec_;
-
-    // Calculate the earlybird percentage to be applied
-    int128_t ebPercentage_ = weeks_ * bonusPerc;
-
-    // Calcluate the bonus percentage to apply up to percentage_precision number of digits
-    int128_t bonusPercentage_ = (precision_ * percentage_precision);
-    bonusPercentage_ += ebPercentage_;
-    bonusPercentage_ /= percentage_precision;
-
-    // Calculate the bonus percentage for the issuer
-    int128_t issuerPercentage_(issuerPerc);
-    issuerPercentage_ *= precision_;
-    issuerPercentage_ /= percentage_precision;
-
-    // Precision for bitcoin amounts (satoshi)
-    int128_t satoshi_precision_(100000000L);
-
-    // Total tokens including remainders
-    cpp_int createdTokens(amtTransfer);
-    createdTokens *= cpp_int(numProps);
-    createdTokens *= cpp_int(bonusPercentage_);
-
-    cpp_int issuerTokens = createdTokens / satoshi_precision_;
-    issuerTokens /= precision_;
-    issuerTokens *= (issuerPercentage_ / 100);
-    issuerTokens *= precision_;
-
-    cpp_int createdTokens_int = createdTokens / precision_;
-    createdTokens_int /= satoshi_precision_;
-
-    cpp_int issuerTokens_int = issuerTokens / precision_;
-    issuerTokens_int /= satoshi_precision_;
-    issuerTokens_int /= 100;
-
-    cpp_int newTotalCreated = totalTokens + createdTokens_int + issuerTokens_int;
-
-    if (newTotalCreated > MAX_INT_8_BYTES) {
-        cpp_int maxCreatable = MAX_INT_8_BYTES - totalTokens;
-        cpp_int created = createdTokens_int + issuerTokens_int;
-
-        // Calcluate the ratio of tokens for what we can create and apply it
-        cpp_int ratio = created * precision_;
-        ratio *= satoshi_precision_;
-        ratio /= maxCreatable;
-
-        // The tokens for the issuer
-        issuerTokens_int = issuerTokens_int * precision_;
-        issuerTokens_int *= satoshi_precision_;
-        issuerTokens_int /= ratio;
-
-        // The tokens for the user
-        createdTokens_int = MAX_INT_8_BYTES - issuerTokens_int;
-
-        // Close the crowdsale after assigning all tokens
-        close_crowdsale = true;
-    }
-
-    // The tokens to credit
-    assert(createdTokens_int <= std::numeric_limits<int64_t>::max());
-    assert(issuerTokens_int <= std::numeric_limits<int64_t>::max());
-    tokens = std::make_pair(createdTokens_int.convert_to<int64_t>(), issuerTokens_int.convert_to<int64_t>());
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -4222,199 +4137,5 @@ const std::vector<unsigned char> GetOmMarker()
     static unsigned char pch[] = {0x6f, 0x6d, 0x6e, 0x69}; // Hex-encoded: "omni"
 
     return std::vector<unsigned char>(pch, pch + sizeof(pch) / sizeof(pch[0]));
-}
-
- // the 31-byte packet & the packet #
- // int interpretPacket(int blocknow, unsigned char pkt[], int size)
- //
- // RETURNS:  0 if the packet is fully valid
- // RETURNS: <0 if the packet is invalid
- // RETURNS: >0 the only known case today is: return PKT_RETURNED_OBJECT
- //
- // 
- // the following functions may augment the amount in question (nValue):
- // DEx_offerCreate()
- // DEx_offerUpdate()
- // DEx_acceptCreate()
- // DEx_payment() -- DOES NOT fit nicely into the model, as it currently is NOT a MP TX (not even in leveldb) -- gotta rethink
- //
- // optional: provide the pointer to the CMPOffer object, it will get filled in
- // verify that it does via if (MSC_TYPE_TRADE_OFFER == mp_obj.getType())
- //
-int CMPTransaction::interpretPacket()
-{
-    if (rpcOnly) {
-        PrintToLog("%s(): ERROR: attempt to execute logic in RPC mode\n", __func__);
-        return (PKT_ERROR -1);
-    }
-
-    if (!interpret_Transaction()) {
-        return (PKT_ERROR -2);
-    }
-
-    LOCK(cs_tally);
-
-    switch (type) {
-        case MSC_TYPE_SIMPLE_SEND:
-            return logicMath_SimpleSend();
-
-        case MSC_TYPE_SEND_TO_OWNERS:
-            return logicMath_SendToOwners();
-
-        case MSC_TYPE_TRADE_OFFER:
-            return logicMath_TradeOffer();
-
-        case MSC_TYPE_ACCEPT_OFFER_BTC:
-            return logicMath_AcceptOffer_BTC();
-
-        case MSC_TYPE_METADEX_TRADE:
-            return logicMath_MetaDExTrade();
-
-        case MSC_TYPE_METADEX_CANCEL_PRICE:
-            return logicMath_MetaDExCancelPrice();
-
-        case MSC_TYPE_METADEX_CANCEL_PAIR:
-            return logicMath_MetaDExCancelPair();
-
-        case MSC_TYPE_METADEX_CANCEL_ECOSYSTEM:
-            return logicMath_MetaDExCancelEcosystem();
-
-        case MSC_TYPE_CREATE_PROPERTY_FIXED:
-            return logicMath_CreatePropertyFixed();
-
-        case MSC_TYPE_CREATE_PROPERTY_VARIABLE:
-            return logicMath_CreatePropertyVariable();
-
-        case MSC_TYPE_CLOSE_CROWDSALE:
-            return logicMath_CloseCrowdsale();
-
-        case MSC_TYPE_CREATE_PROPERTY_MANUAL:
-            return logicMath_CreatePropertyManaged();
-
-        case MSC_TYPE_GRANT_PROPERTY_TOKENS:
-            return logicMath_GrantTokens();
-
-        case MSC_TYPE_REVOKE_PROPERTY_TOKENS:
-            return logicMath_RevokeTokens();
-
-        case MSC_TYPE_CHANGE_ISSUER_ADDRESS:
-            return logicMath_ChangeIssuer();
-
-        case OMNICORE_MESSAGE_TYPE_ALERT:
-            return logicMath_Alert();
-    }
-
-    return (PKT_ERROR -100);
-}
-
-int CMPTransaction::logicMath_SimpleSend()
-{
-    if (!IsTransactionTypeAllowed(block, property, type, version)) {
-        PrintToLog("%s(): rejected: type %d or version %d not permitted for property %d at block %d\n",
-                __func__,
-                type,
-                version,
-                property,
-                block);
-        return (PKT_ERROR_SEND -22);
-    }
-
-    if (nValue <= 0 || MAX_INT_8_BYTES < nValue) {
-        PrintToLog("%s(): rejected: value out of range or zero: %d", __func__, nValue);
-        return (PKT_ERROR_SEND -23);
-    }
-
-    if (!_my_sps->hasSP(property)) {
-        PrintToLog("%s(): rejected: property %d does not exist\n", __func__, property);
-        return (PKT_ERROR_SEND -24);
-    }
-
-    int64_t nBalance = getMPbalance(sender, property, BALANCE);
-    if (nBalance < (int64_t) nValue) {
-        PrintToLog("%s(): rejected: sender %s has insufficient balance of property %d [%s < %s]\n",
-                __func__,
-                sender,
-                property,
-                FormatMP(property, nBalance),
-                FormatMP(property, nValue));
-        return (PKT_ERROR_SEND -25);
-    }
-
-    // ------------------------------------------
-
-    // Special case: if can't find the receiver -- assume send to self!
-    if (receiver.empty()) {
-        receiver = sender;
-    }
-
-    // Move the tokens
-    assert(update_tally_map(sender, property, -nValue, BALANCE));
-    assert(update_tally_map(receiver, property, nValue, BALANCE));
-
-    // Is there an active crowdsale running from this recepient?
-    {
-        CMPCrowd* pcrowdsale = getCrowd(receiver);
-
-        if (pcrowdsale && pcrowdsale->getCurrDes() == property) {
-            CMPSPInfo::Entry sp;
-            bool spFound = _my_sps->getSP(pcrowdsale->getPropertyId(), sp);
-
-            PrintToLog("INVESTMENT SEND to Crowdsale Issuer: %s\n", receiver);
-
-            if (spFound) {
-                // Holds the tokens to be credited to the sender and receiver
-                std::pair<int64_t, int64_t> tokens;
-
-                // Passed by reference to determine, if max_tokens has been reached
-                bool close_crowdsale = false;
-
-                // Units going into the calculateFundraiser function must
-                // match the unit of the fundraiser's property_type.
-                // By default this means satoshis in and satoshis out.
-                // In the condition that the fundraiser is divisible,
-                // but indivisible tokens are accepted, it must account for
-                // 1.0 Div != 1 Indiv, but actually 1.0 Div == 100000000 Indiv.
-                // The unit must be shifted or the values will be incorrect,
-                // which is what is checked below.
-                if (!isPropertyDivisible(property)) {
-                    nValue = nValue * 1e8;
-                }
-
-                // Calculate the amounts to credit for this fundraiser
-                calculateFundraiser(nValue, sp.early_bird, sp.deadline, blockTime,
-                        sp.num_tokens, sp.percentage, getTotalTokens(pcrowdsale->getPropertyId()),
-                        tokens, close_crowdsale);
-
-                if (msc_debug_sp) {
-                    PrintToLog("%s(): granting via crowdsale to user: %s %d (%s)\n",
-                            __func__, FormatMP(property, tokens.first), property, strMPProperty(property));
-                    PrintToLog("%s(): granting via crowdsale to issuer: %s %d (%s)\n",
-                            __func__, FormatMP(property, tokens.second), property, strMPProperty(property));
-                }
-
-                // Update the crowdsale object
-                pcrowdsale->incTokensUserCreated(tokens.first);
-                pcrowdsale->incTokensIssuerCreated(tokens.second);
-
-                // Data to pass to txFundraiserData
-                int64_t txdata[] = {(int64_t) nValue, blockTime, tokens.first, tokens.second};
-                std::vector<int64_t> txDataVec(txdata, txdata + sizeof(txdata) / sizeof(txdata[0]));
-
-                // Insert data about crowdsale participation
-                pcrowdsale->insertDatabase(txid, txDataVec);
-
-                // Credit tokens for this fundraiser
-                update_tally_map(sender, pcrowdsale->getPropertyId(), tokens.first, BALANCE);
-                update_tally_map(receiver, pcrowdsale->getPropertyId(), tokens.second, BALANCE);
-
-                // Close crowdsale, if we hit MAX_TOKENS
-                if (close_crowdsale) {
-                    eraseMaxedCrowdsale(receiver, blockTime, block);
-                }
-            }
-        }
-    }
-
-    return 0;
 }
