+++ src/omnicore/sp.h
@@ -20,6 +20,7 @@ class uint256;
 #include <fstream>
 #include <map>
 #include <string>
+#include <utility>
 #include <vector>
 
 /** LevelDB based storage for currencies, smart properties and tokens.
@@ -222,6 +223,10 @@ bool isCrowdsalePurchase(const uint256& txid, const std::string& address, int64_
 int64_t calculateFractional(uint16_t propType, uint8_t bonusPerc, int64_t fundraiserSecs,
         int64_t numProps, uint8_t issuerPerc, const std::map<uint256, std::vector<int64_t> >& txFundraiserData,
         const int64_t amountPremined);
+
+void calculateFundraiser(int64_t amtTransfer, uint8_t bonusPerc,
+        int64_t fundraiserSecs, int64_t currentSecs, int64_t numProps, uint8_t issuerPerc, int64_t totalTokens,
+        std::pair<int64_t, int64_t>& tokens, bool& close_crowdsale);
 
 void eraseMaxedCrowdsale(const std::string& address, int64_t blockTime, int block);
